% Generated by roxygen2 (4.0.1): do not edit by hand
\name{data.file.exists}
\alias{data.file.exists}
\title{Check whether a file exists in the data folder}
\usage{
data.file.exists(file, data_folder = NULL)
}
\arguments{
\item{file:}{the file name to save (relative to data_folder)}

\item{data_folder:}{the data folder, defaults to get.data.folder()}
}
\description{
Check whether a file exists in the data folder
}

