% Generated by roxygen2 (4.0.2): do not edit by hand
\name{data.load}
\alias{data.load}
\title{Load a file from the data folder}
\usage{
data.load(fn, data_folder = NULL)
}
\arguments{
\item{fn:}{the file name, including path (relative to data_folder)}

\item{data_folder:}{the data folder, defaults to data.folder()}
}
\description{
Load a file from the data folder
}

