% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_describe_task}
\alias{datasync_describe_task}
\title{Returns metadata about a task}
\usage{
datasync_describe_task(TaskArn)
}
\arguments{
\item{TaskArn}{[required] The Amazon Resource Name (ARN) of the task to describe.}
}
\description{
Returns metadata about a task.
}
\section{Request syntax}{
\preformatted{datasync$describe_task(
  TaskArn = "string"
)
}
}

\keyword{internal}
