% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasync_operations.R
\name{datasync_describe_task_execution}
\alias{datasync_describe_task_execution}
\title{Returns detailed metadata about a task that is being executed}
\usage{
datasync_describe_task_execution(TaskExecutionArn)
}
\arguments{
\item{TaskExecutionArn}{[required] The Amazon Resource Name (ARN) of the task that is being executed.}
}
\description{
Returns detailed metadata about a task that is being executed.
}
\section{Request syntax}{
\preformatted{datasync$describe_task_execution(
  TaskExecutionArn = "string"
)
}
}

\keyword{internal}
