% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.data.folder}
\alias{get.data.folder}
\title{Get the data folder from the R_DATA_FOLDER environment variable, defaulting to ~/data}
\usage{
get.data.folder()
}
\description{
Get the data folder from the R_DATA_FOLDER environment variable, defaulting to ~/data
}

