% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load.data}
\alias{load.data}
\title{Load a file from the data folder}
\usage{
load.data(file, data_folder = NULL, envir = .GlobalEnv)
}
\arguments{
\item{file:}{the file name, including path (relative to data_folder)}

\item{data_folder:}{the data folder, defauts to get.data.folder()}

\item{envir:}{the environment to load the data into, defaults to the global environment}
}
\description{
Load a file from the data folder
}

