% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rsync}
\alias{rsync}
\title{Conduct an rsync with a remote host}
\usage{
rsync(remote_host, remote_folder = NULL, local_folder = NULL, dry_run = F)
}
\arguments{
\item{remote_host:}{the hostname of the remote host to sync with}

\item{remote_folder:}{the remote folder to sync with (defaults to data.folder())}

\item{local_folder:}{the local folder to sync (defaults to data.folder())}

\item{dry_run:}{if TRUE, don't actually do the syncing, just check what would be done (i.e. rsync --dry-run)}
}
\description{
Uses a pair of system calls to rsync, so rsync needs to be installed and the current user
needs to have credentials to login to the remote host automatically.
}

