% Generated by roxygen2 (4.0.2): do not edit by hand
\name{save.data}
\alias{save.data}
\title{Save variables into a file in the data folder}
\usage{
save.data(..., file, data_folder = NULL)
}
\arguments{
\item{...:}{the variables to save}

\item{file:}{the file name to save (relative to data_folder)}

\item{data_folder:}{the data folder, defaults to get.data.folder()}
}
\description{
Save variables into a file in the data folder
}

